Renamet useVedtaksperioder -> useVedtakSteg fordi contexten inneholder mye mer enn bare vedtaksperioder
