+++ src/frontend/komponenter/Fagsak/Vedtak/OppsummeringVedtak.tsx
@@ -11,7 +11,7 @@ import { Vedtaksalert } from './Vedtaksalert';
 import { Vedtaksbrev } from './Vedtaksbrev';
 import Vedtaksmeny from './Vedtaksmeny';
 import { useBehandling } from '../../../context/behandlingContext/BehandlingContext';
-import { useVedtaksperioder } from '../../../context/behandlingContext/useVedtaksperioder';
+import { useVedtakSteg } from '../../../context/behandlingContext/useVedtakSteg';
 import { useSimulering } from '../../../context/SimuleringContext';
 import useSakOgBehandlingParams from '../../../hooks/useSakOgBehandlingParams';
 import type { IBehandling } from '../../../typer/behandling';
@@ -48,7 +48,7 @@ const OppsummeringVedtak: React.FunctionComponent<IVedtakProps> = ({ åpenBehand
         erUlagretNyFeilutbetaltValutaPeriode,
         erUlagretNyRefusjonEøsPeriode,
         sammensattKontrollsakContext,
-    } = useVedtaksperioder();
+    } = useVedtakSteg();
 
     const { behandlingErMigreringMedAvvikUtenforBeløpsgrenser } = useSimulering();
